revise local bin algorithm to meet specification

+++ Library/util.h
@@ -7,7 +7,7 @@ using namespace std;
 int numberOfDigits(int number);
 
 template <typename T>
-void print(vector<T>& vec, int trim_count = 10);
+void print(vector<T>& vec, size_t trim_count = 10);
 
 template<typename T>
 void readNumber(T& number, string prompt, T min = numeric_limits<T>::lowest(), T max = numeric_limits<T>::max());
